chore: remove db migration on startup in order to start gunicorn directly
This removes the `flask db upgrade` command that's run on every new
deployment, thereby requiring us to perform db migrations manually now.

We could also consider doing this as part of the deployment script, but
it would likely require some downtime during redeploys.

The reasoning is that when started the old way through `/bin/sh`,
signals sent by kubernetes (like SIGTERM) aren't passed along to
gunicorn, and the server will not terminate as expected.

This was likely also the cause of db connection issues reported by Sentry
during shutdown - the proxy terminates, but the gunicorn server is
still running and queried by the health check service.